indexer-alt: factor out DB connection stats collector
## Description

...and use it in the RPC implementation as well. A parameter has been
added to add a prefix to all metric names. This ensures that if an
Indexer and RPC service are sharing a metrics service, they won't stomp
on each others metrics.

## Test plan

Run the RPC service and check its metrics.
