+++ crates/sui-indexer-alt-framework/src/metrics.rs
@@ -4,14 +4,11 @@
 use std::sync::{atomic::AtomicU64, Arc};
 
 use prometheus::{
-    core::{Collector, Desc},
-    proto::{Counter, Gauge, LabelPair, Metric, MetricFamily, MetricType, Summary},
     register_histogram_vec_with_registry, register_histogram_with_registry,
     register_int_counter_vec_with_registry, register_int_counter_with_registry,
     register_int_gauge_vec_with_registry, register_int_gauge_with_registry, Histogram,
     HistogramVec, IntCounter, IntCounterVec, IntGauge, IntGaugeVec, Registry,
 };
-use sui_pg_db::Db;
 use tracing::warn;
 
 use crate::{ingestion::error::Error, pipeline::Processor};
@@ -132,12 +129,6 @@ pub(crate) struct IndexerMetrics {
     pub watermark_timestamp_in_db_ms: IntGaugeVec,
     pub watermark_reader_lo_in_db: IntGaugeVec,
     pub watermark_pruner_hi_in_db: IntGaugeVec,
-}
-
-/// Collects information about the database connection pool.
-pub(crate) struct DbConnectionStatsCollector {
-    db: Db,
-    desc: Vec<(MetricType, Desc)>,
 }
 
 /// A helper struct to report metrics regarding the checkpoint lag at various points in the indexer.
@@ -614,100 +605,6 @@ impl IndexerMetrics {
     }
 }
 
-impl DbConnectionStatsCollector {
-    pub(crate) fn new(db: Db) -> Self {
-        let desc = vec![
-            (
-                MetricType::GAUGE,
-                desc(
-                    "db_connections",
-                    "Number of connections currently being managed by the pool",
-                ),
-            ),
-            (
-                MetricType::GAUGE,
-                desc(
-                    "db_idle_connections",
-                    "Number of idle connections in the pool",
-                ),
-            ),
-            (
-                MetricType::COUNTER,
-                desc("db_connect_direct", "Connections that did not have to wait"),
-            ),
-            (
-                MetricType::SUMMARY,
-                desc("db_connect_waited", "Connections that had to wait"),
-            ),
-            (
-                MetricType::COUNTER,
-                desc(
-                    "db_connect_timed_out",
-                    "Connections that timed out waiting for a connection",
-                ),
-            ),
-            (
-                MetricType::COUNTER,
-                desc(
-                    "db_connections_created",
-                    "Connections that have been created in the pool",
-                ),
-            ),
-            (
-                MetricType::COUNTER,
-                desc_with_labels(
-                    "db_connections_closed",
-                    "Total connections that were closed",
-                    &["reason"],
-                ),
-            ),
-        ];
-
-        Self { db, desc }
-    }
-}
-
-impl Collector for DbConnectionStatsCollector {
-    fn desc(&self) -> Vec<&Desc> {
-        self.desc.iter().map(|d| &d.1).collect()
-    }
-
-    fn collect(&self) -> Vec<MetricFamily> {
-        let state = self.db.state();
-        let stats = state.statistics;
-
-        vec![
-            gauge(&self.desc[0].1, state.connections as f64),
-            gauge(&self.desc[1].1, state.idle_connections as f64),
-            counter(&self.desc[2].1, stats.get_direct as f64),
-            summary(
-                &self.desc[3].1,
-                stats.get_wait_time.as_millis() as f64,
-                stats.get_waited + stats.get_timed_out,
-            ),
-            counter(&self.desc[4].1, stats.get_timed_out as f64),
-            counter(&self.desc[5].1, stats.connections_created as f64),
-            counter_with_labels(
-                &self.desc[6].1,
-                &[
-                    ("reason", "broken", stats.connections_closed_broken as f64),
-                    ("reason", "invalid", stats.connections_closed_invalid as f64),
-                    (
-                        "reason",
-                        "max_lifetime",
-                        stats.connections_closed_max_lifetime as f64,
-                    ),
-                    (
-                        "reason",
-                        "idle_timeout",
-                        stats.connections_closed_idle_timeout as f64,
-                    ),
-                ],
-            ),
-        ]
-    }
-}
-
 impl CheckpointLagMetricReporter {
     pub fn new(
         checkpoint_time_lag_histogram: Histogram,
@@ -748,89 +645,6 @@ impl CheckpointLagMetricReporter {
             self.latest_checkpoint_time_lag_gauge.set(lag);
         }
     }
-}
-
-fn desc(name: &str, help: &str) -> Desc {
-    desc_with_labels(name, help, &[])
-}
-
-fn desc_with_labels(name: &str, help: &str, labels: &[&str]) -> Desc {
-    Desc::new(
-        name.to_string(),
-        help.to_string(),
-        labels.iter().map(|s| s.to_string()).collect(),
-        Default::default(),
-    )
-    .expect("Bad metric description")
-}
-
-fn gauge(desc: &Desc, value: f64) -> MetricFamily {
-    let mut g = Gauge::default();
-    let mut m = Metric::default();
-    let mut mf = MetricFamily::new();
-
-    g.set_value(value);
-    m.set_gauge(g);
-
-    mf.mut_metric().push(m);
-    mf.set_name(desc.fq_name.clone());
-    mf.set_help(desc.help.clone());
-    mf.set_field_type(MetricType::COUNTER);
-    mf
-}
-
-fn counter(desc: &Desc, value: f64) -> MetricFamily {
-    let mut c = Counter::default();
-    let mut m = Metric::default();
-    let mut mf = MetricFamily::new();
-
-    c.set_value(value);
-    m.set_counter(c);
-
-    mf.mut_metric().push(m);
-    mf.set_name(desc.fq_name.clone());
-    mf.set_help(desc.help.clone());
-    mf.set_field_type(MetricType::GAUGE);
-    mf
-}
-
-fn counter_with_labels(desc: &Desc, values: &[(&str, &str, f64)]) -> MetricFamily {
-    let mut mf = MetricFamily::new();
-
-    for (name, label, value) in values {
-        let mut c = Counter::default();
-        let mut l = LabelPair::default();
-        let mut m = Metric::default();
-
-        c.set_value(*value);
-        l.set_name(name.to_string());
-        l.set_value(label.to_string());
-
-        m.set_counter(c);
-        m.mut_label().push(l);
-        mf.mut_metric().push(m);
-    }
-
-    mf.set_name(desc.fq_name.clone());
-    mf.set_help(desc.help.clone());
-    mf.set_field_type(MetricType::COUNTER);
-    mf
-}
-
-fn summary(desc: &Desc, sum: f64, count: u64) -> MetricFamily {
-    let mut s = Summary::default();
-    let mut m = Metric::default();
-    let mut mf = MetricFamily::new();
-
-    s.set_sample_sum(sum);
-    s.set_sample_count(count);
-    m.set_summary(s);
-
-    mf.mut_metric().push(m);
-    mf.set_name(desc.fq_name.clone());
-    mf.set_help(desc.help.clone());
-    mf.set_field_type(MetricType::SUMMARY);
-    mf
 }
 
 #[cfg(test)]
