+++ crates/sui-indexer-alt-metrics/Cargo.toml
@@ -14,3 +14,5 @@ prometheus.workspace = true
 tokio.workspace = true
 tokio-util.workspace = true
 tracing.workspace = true
+
+sui-pg-db.workspace = true
